refactor resources pkg
